requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
click==8.1.6
colorama==0.4.6
commonmark==0.9.1
dnspython==2.4.1
fastapi==0.101.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
motor==3.2.0
psycopg2==2.9.6
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.15.1
pymongo==4.4.1
pyscopg2==66.0.2
python-dotenv==1.0.0
rich==12.6.0
shellingham==1.5.0.post1
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typer==0.7.0
typing_extensions==4.7.1
uvicorn==0.23.2
